Implement filling of empty fields with random data (DM-20580)
Cassandra does not store NULL so for realistic test we need non-NULL
payload for all table columns. I added new configuration option for
Cassandra backend which fills columns that are not explicitly set with
random data. I'm doing it the backend instead of ap_proto because it is
much simplere to do here and I need a quick and dirty way to do it right
now. For final implementation we won't need it ao it will be removed
when we settle on what final implementation is going to look like.