Change schema file discovery to use environment variables and expandvars
This changes the config items to support environment variables rather
than using getPackageDir. This allows runtime evaluation of the
file path rather than config read evaluation (which then gets
written to quantum graph files). This allows graph building to
run on a different system to the system that executes that graph.