WIP: Preparation for updating ApdbCassandra for new interface
Previous commits on this branch come from rebasing of the testing branch
(u/andy-slac/cassandra-2), next step is to update the implementation
for new abstact interface. This commit adds a unit test for
ApdbCassandra to make sure that it imports and can trivially function.
This unit test can only be ran against actual Cassandra cluster, needs
special environment. I added a protection for the case of the missing
cassandra-driver package, it is still possible to import the module in
that case but instantiation of ApdbCassandra fails.