Add methods to support future PPDB replication (DM-33313)
Apdb interface was extended with few methods that will be needed for the
replication to/from PPDB. SQL implemention has these new methods working
but Cassandra implementation will be done on a separate ticket (it needs
schema extension and some experimenting). This new interface may still
change when the replication system implementation starts.