Implement ApdbSql.containsCcdVisit.
This method does the same thing as Apdb.containsVisitDetector, but
takes as input a "packed" visit-detector ID instead of separate visit
and detector IDs.

This alternative is necessary because, at present, the packed ID is the
only form of the information in the database, and the packing process
requires information that dax.apdb.Apdb should not have access to.
containsCcdVisit can be replaced with contains once either a dedicated
table is created, or Dia*Source use visit and detector internally.