Add support for version checks to SQL backend.
Apdb class now has an associated version number and schema version number
(latter comes from YAML schema). Both version numbers are stored in
metadata table when schema is initialized. Versions are read when
ApdbSql is instantiated and version numbers are compared to what is
defined in YAML and in the code.