Change `Apdb.makeSchema` to be class method.
Having `makeSchema` as an instance method causes too many issues,
as the state of the instance has to be reset after schema is created
or re-created. Now Apdb instances will be created only on populated
database schema. Base class method can be used to call `makeSchema`
and it will dispatch to a correct sub-class based on confguration
object type.

This change means that in-memory SQLite database cannot be used with
ApdbSql, in-memory database was only useful for testing.