Change granularity of DiainsertId.
DiaInsertId structure changed, `id` now has integer type. One DiaInsertId
corresponds to multiple `store` calls. This is needed to reduce the number
of transfer chunks. One chunk corresponds to a configurable time window (10
minutes by default). Note that this update changes version number for replica
classes to 1.0.0. We should not have databases with replication enabled so
this should not be an issue.

Change types of collections returned from ApdbTableData