Add support for APDB index.
Index is a configuration file which maps labels to the URIs of APDB
configuration files. Apdb can be instantiated via `from_uri` method
using `label:` prefix with the name of the label defined in the index file.

CLI sub-command 'list-index` is added to dump the content of the index file.