Handle the case when sqlite database was removed (DM-45683)
SQLite by default creates an empty database when database file given
in URL does not exist. It needs extra convincing with special URL
syntax to not do that. When new APDB sqlite instance is created, its
config file will have an updated URL with that extra syntax. When old
config file is read which does not have that syntax, we modify URL and
add the same options.