Fix handling of timezones for input timestamps.
If Postgres server timezone is seto to something then UTC we
need to add UTC timezone to input timestamps so that server
does not try to convert them to server timezone. I also added
check for input timestamps with timezones, and convert them
to UTC as well (this may not be necessary, just to be safe).