Improve timestamp handling for Cassandra, fixed truncation bug.
If input timestamps are aware thjen convert them to UTC and drop
timezone. Cassandra does not need timezone info if input is in UTC.
On return Cassandra makes naive time (in UTC), so no conversion is
necessary. There was a bug in converting timestamps to milliseconds
that dropped fractional part, this is now fixed.