Add options to limit the size of batch queries in Cassandra (DM-48391)
ApdbCassandra configuration adds options to limit the number of rows in
batch queries and the size of the batch queries. This includes fixing
the error related to 64k limit of the batch queries for updates of
DiaObjectLastToPartition table. There are still some batch queries that
do not check the limits, but those are not likely to exceed Cassandra
limits.