Add extra table to track visit/detector (DM-51262)
Previous attempts to solve performance issue of `containsVisitDetector`
method were not successful. This patch avoids long queries on regular
tables by introducing a special table that only tracks visits and detectors.
The version of ApdbCassandra class upgraded to 0.1.2, and it stays compatible
with the previous version by using old inefficient queries when the table
is missing.

Automatic table cleanup is is implemented by setting TTL option to one day.