Add _storeUpdateRecord to Apdb and subclasses (DM-50190)
This method will be used by the methods that update existing
records (methods will be added on a new ticket). Update records
are stored in a separate table which is partitioned similarly to
other replica tables. ApdbReplica subclasses implement
`getUpdateRecordChunks()` method to retrieve records for
a number of replica chunks.