Add support for DiaObjectLast.validityStart (DM-52827)
APDB schema version 9.1.0 in sdm_schemas adds validityStartMjdTai column
to DiaObjectLast table. This column needs to be filled by Apdb classes
similarly to how it is done for DiaObject column. This patch adds needed
fuctionality and it can also handle the case of older schema and will
aviod filling the column if it is missing from database schema.
Correspondingly version of ApdCassandra/ApdbSql classes increments
minor version.

Migration scripts for the new schema version is added on dax_apdb_migrate.