Remove SSObject table from API and schema (DM-53385)
Apdb had methods for query/insert into SSObject table that were never
used by anyhting. The backends created the table in database to support
the API. Modern idea is that we do not need that table in APDB, so I
am removing both the methods and instantiation of SSObject table.